refactor(test): mirror CLI entry and init test paths
