+++ packages/pyric-tools/test/cli/init.test.ts
